+++ __tests__/serverTests.test.js
@@ -3,10 +3,9 @@
  */
 
 import React from 'react';
-
 import ReactDOMServer from 'react-dom/server';
 import ParallaxProvider from 'components/ParallaxProvider';
-import ParallaxController from 'modules/ParallaxController';
+import ParallaxController from 'classes/ParallaxController';
 
 describe('Expect', () => {
     it('<ParallaxProvider> to not init the controller on the server', () => {
