+++ src/helpers/index.js
@@ -1,0 +1,7 @@
+export ParallaxContext from './ParallaxContext';
+export { percentMoved } from './percentMoved';
+export { setParallaxStyles, resetStyles } from './elementStyles';
+export { getParallaxOffsets } from './getParallaxOffsets';
+export { isElementInView } from './isElementInView';
+export getOffsets from './getOffsets';
+export getCache from './getCache';
