+++ src/helpers/percentMoved.js
