+++ src/index.js
@@ -2,4 +2,4 @@ export withController from './components/withController';
 export Parallax from './components/Parallax';
 export ParallaxProvider from './components/ParallaxProvider';
 export ParallaxBanner from './components/ParallaxBanner';
-export ParallaxController from './modules/ParallaxController';
+export ParallaxController from './classes/ParallaxController';
