+++ src/utils/index.js
@@ -1,11 +1,3 @@
-export { percentMoved } from './percentMoved';
-export { setParallaxStyles, resetStyles } from './elementStyles';
-export { getParallaxOffsets } from './getParallaxOffsets';
-export { isElementInView } from './isElementInView';
-
-export getOffsets from './getOffsets';
-export getCache from './getCache';
-
 export parseValueAndUnit from './parseValueAndUnit';
 export scaleBetween from './scaleBetween';
 export testForPassiveScroll from './testForPassiveScroll';
